+++ Remnants/Main.cs
@@ -7,7 +7,7 @@ using Remnants.Data;
 
 namespace Remnants
 {
-    [BepInDependency("evaisa.lethallib", "0.16.0")]
+    [BepInDependency("evaisa.lethallib", "0.16.1")]
     [BepInDependency("ainavt.lc.LethalConfig", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInPlugin(modGUID, modName, modVersion)]
 
@@ -16,7 +16,7 @@ namespace Remnants
         #region Variables
         private const string modGUID = "KawaiiBone.Remnants";
         private const string modName = "Remnants";
-        private const string modVersion = "1.3.8";
+        private const string modVersion = "1.3.9";
 
         public static Remnants Instance;
         private readonly Harmony _harmony = new Harmony(modGUID);
@@ -29,7 +29,7 @@ namespace Remnants
         public SpawnBodiesBehaviour SpawningBodyBeh = new SpawnBodiesBehaviour();
         public RemnantItemsBehaviour RemnantItemsBeh = new RemnantItemsBehaviour();
         public RandomizeBatteriesBehaviour ItemsBatteriesBeh = new RandomizeBatteriesBehaviour();
-        public RemnantItemsLocationsBehaviour RegisterItemLocationsBeh = new RemnantItemsLocationsBehaviour();
+        public GrabbableObjsSpawnListBehaviour GrabbableObjsSpawnListBeh = new GrabbableObjsSpawnListBehaviour();
         public SpawnRemnantItemsBehaviour SpawnRemnantItemsBeh = new SpawnRemnantItemsBehaviour();
 
         private RegisterItemsBehaviour _registerItemsBehaviour = new RegisterItemsBehaviour();
@@ -58,9 +58,11 @@ namespace Remnants
             _harmony.PatchAll(typeof(BodySuitBehaviour));
             _harmony.PatchAll(typeof(RegisterSuitsPatch));
             _harmony.PatchAll(typeof(AddRemnantItemsToItemList));
-            _harmony.PatchAll(typeof(TerminalScanItemsPatch));
+            _harmony.PatchAll(typeof(TerminalScanItemsTranspiler));
+            _harmony.PatchAll(typeof(EndRoundStatsPatch));
             _harmony.PatchAll(typeof(BeltBagTranspiler));
             _harmony.PatchAll(typeof(SaveGameTranspiler));
+            _harmony.PatchAll(typeof(GrabbableObjToListStartPatch));
             _harmony.PatchAll(typeof(Remnants));
             _registerItemsBehaviour.Initialize();
             RegisterBodiesSpawn.Initialize();
@@ -68,7 +70,7 @@ namespace Remnants
             RegisterBodySuits.Initialize();
             SpawningBodyBeh.Initialize();
             ItemsBatteriesBeh.Initialize();
-            RegisterItemLocationsBeh.Initialize();
+            GrabbableObjsSpawnListBeh.Initialize();
             SpawnRemnantItemsBeh.Initialize();
             Mls.LogInfo("modGUID has loaded");
         }